File restructure
* Move CommonJS code into /cjs folder, ESM into /esm folder
* Update tests + package.json in repo root
