+++ js/src/cjs/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  Derek: require("./derek").Derek,
+  Parser: require("./parse/").Parser,
+};
