+++ js/src/cjs/parse/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  Parser: require("./Parser").Parser,
+};
