+++ js/src/esm/index.js
@@ -1,0 +1,2 @@
+import { Derek, Parser } from "../cjs";
+export { Derek, Parser };
