--- js/src/index.cjs
@@ -1,4 +1,0 @@
-module.exports = {
-  Derek: require("./derek.cjs").Derek,
-  Parser: require("./parse/index.cjs").Parser,
-};
