--- js/src/parse/index.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-  Parser: require("./Parser.cjs").Parser,
-};
