--- js/src/wrapper.mjs
@@ -1,2 +1,0 @@
-import { Derek, Parser } from "./index.cjs";
-export { Derek, Parser };
