+++ js/tests/cjs/_parse/oas3.test.js
@@ -1,5 +1,5 @@
-const { Derek } = require("../../src/derek.cjs");
-const { Parser } = require("../../src/parse/parser.cjs");
+const { Derek } = require("../../../src/cjs/derek");
+const { Parser } = require("../../../src/cjs/parse/parser");
 
 function schema(obj, ...kwargs) {
   const node = Derek.tree(obj, ...kwargs);
